# Use naïve algorithm for generic A[tc]_mul_B!